Tests: Block Editor: Form Setting: Refactor Tests
